Moved last general monster things from dragon
ABILITIES, SKILLS, plus a couple math functions.
